Use int instead of byte for MinStarboardReactions

+++ src/BChan.Bot/Infra/BotConfigurationManager.cs
@@ -30,7 +30,7 @@ public sealed class BotConfigurationManager(AppDbContext dbContext)
 	public async Task RemoveAutoRole(ulong id, CancellationToken token) =>
 		await EditAndSave(config => config.AutoRolesIds.Remove(id), token);
 
-	public async Task<byte> GetMinStarboardReactions(CancellationToken token) =>
+	public async Task<int> GetMinStarboardReactions(CancellationToken token) =>
 		(await GetBotConfiguration(token)).MinStarboardReactions;
 
 	public async Task SetMinStarboardReactions(byte value, CancellationToken token) =>
